build(nix): remove flake-parts and simplify flake
We're using a very small part of flake-parts up until now, all of which
can be achieved with the include "forAllSystems" method. This commit
removes the dependency on flake-parts which should speed up evaluation
slightly, and make the flake a little easier to parse.

Note that as part of this change, there is also now an overlay provided
which people can use in their system configurations.