Nix flake tweaks (#69)
* chore: format flake.nix with nixpkgs-fmt

* build: include 'result' in .gitignore for nix build output

* refactor(nix): split package into its own file

* build(nix): remove flake-parts and simplify flake

We're using a very small part of flake-parts up until now, all of which
can be achieved with the include "forAllSystems" method. This commit
removes the dependency on flake-parts which should speed up evaluation
slightly, and make the flake a little easier to parse.

Note that as part of this change, there is also now an overlay provided
which people can use in their system configurations.

* build(nix): include 'meta' section in magmawm package

* build(nix): rust-overlay's nixpkgs follows this flake's nixpkgs

* ci: add workflow to validate nix builds